requirements.txt
tkinter
opencv-python
Pillow
ultralytics
